create GUI on working
