+++ SudokuSolve.py
@@ -18,9 +18,8 @@ class SudokuCell():
 				type(value) is not int):
 					value = 0
 		
-				
 		if self.is_fixed == True:
-			print("It's fixed cell")
+			#print("It's fixed cell")
 			return 1
 		elif value == 0:
 			return 2
@@ -37,6 +36,8 @@ class SudokuCell():
 		
 		count = 0
 		tmp_value = 0
+		if self.is_fixed:
+			return
 		for i in range(9):
 			if self.judge[i] == True:
 				count += 1
@@ -45,21 +46,28 @@ class SudokuCell():
 			self.SetValue(tmp_value + 1)
 			return 0
 
+
+	def DelJudge(self, index):
+		
+		self.judge[index] = False
+	
+		if sum(self.judge) == 0:
+			print('all candidate delited.')
+		
 class SudokuField():
 	
 	def __init__(self):
 		print('set field')
 		#self.raws = [[None] * 9] * 9
 		self.raws =[]
+		self.cols = []
+		self.areas = []
+		
 		for i in range(9):
 			raw = []
 			for j in range(9):
 				raw.append(SudokuCell())
-			
 			self.raws.append(raw)
-			
-		self.cols = []
-		self.areas = []
 		
 		for i in range(9):
 			col = []
@@ -79,7 +87,12 @@ class SudokuField():
 			for j in range(9):
 				self.raws[i][j].SetValue(i * 9 + j + 1)
 		"""
+	def CutOtherJudge(self, block, no, value):
 		
+		for i in range(9):
+			
+			block[i]
+	
 	def CutJudge(self, raw, col):
 		
 		if self.raws[raw][col].value != ' ':
@@ -89,9 +102,12 @@ class SudokuField():
 		
 		#if idx != ' ':
 			for i in range(9):
-				self.raws[raw][i].judge[idx] = False
-				self.cols[col][i].judge[idx] = False
-				self.areas[area][i].judge[idx] = False
+				if i !=col:
+					self.raws[raw][i].DelJudge(idx)
+				if i != raw:
+					self.cols[col][i].DelJudge(idx)
+				if i != add:
+					self.areas[area][i].DelJudge(idx)
 	
 	
 	def CutJudgeAll(self):
@@ -111,27 +127,58 @@ class SudokuField():
 		#print('{} cell fixed'.format(count))
 		return count
 		
-	
+		
+	def ScanBlock(self, block):
+		
+		for value in range(9):
+			count= 0
+			tmp_add = 0
+			for add in range(9):
+				if block[add].judge[value] == True:
+					count += 1
+					tmp_add = add
+			if (count == 1 and
+					block[tmp_add].is_fixed == False):
+				"""
+				print('add:{} value:{}'.format(tmp_add,value))
+				for i in range(9):
+					print('block {}:'.format(i), end='')
+					for j in range(9):
+						if block[i].judge[j]:
+							print('t',end='')
+						else:
+							print('.',end='')
+					print('')
+				"""
+				block[tmp_add].SetValue(value + 1)
+		
+		
 	def Solve(self):
 		check = 1
 		
-		while check > 0:
-		#for i in range(10):
+		#while check > 0:
+		for i in range(10):
 			self.CutJudgeAll()
+			for i in range(9):
+				self.ScanBlock(self.raws[i])
+				self.ScanBlock(self.cols[i])
+				self.ScanBlock(self.areas[i])
+				
 			check = self.CheckAll()
-			print('{} cell fixed'.format(check))
+			#print('{} cell fixed'.format(check))
+			#self.ShowWithCandidate()
 		
 		
 	def SetTask(self):
-		task = [[4,0,7, 0,1,0, 2,0,9],
-				[0,1,0, 6,0,7, 0,8,0],
-				[3,0,0, 0,0,0, 0,0,0],
-				[0,4,0, 0,0,0, 0,1,0],
-				[7,0,0, 0,0,0, 0,0,3],
-				[0,3,0, 0,0,0, 0,5,0],
-				[0,0,0, 0,0,0, 0,0,7],
-				[0,5,0, 8,0,2, 0,6,0],
-				[1,0,6, 0,9,0, 8,0,2]]
+		task =	[[0,0,0, 0,0,0, 9,0,0],
+						 [0,0,0, 4,9,0, 0,0,6],
+						 [0,9,6, 0,0,0, 0,4,0],
+						 [0,0,0, 0,8,0, 7,0,0],
+						 [0,2,0, 0,3,7, 5,1,0],
+						 [5,0,0, 0,2,0, 3,8,0],
+						 [9,6,0, 1,0,0, 4,0,0],
+						 [8,5,4, 0,7,0, 0,2,1],
+						 [0,3,0, 0,0,0, 8,0,0]]
 						
 		for i in range(9):
 			for j in range(9):
@@ -149,6 +196,29 @@ class SudokuField():
 			print('')
 			if raw % 3 == 2 and raw != 8:
 				print('---+---+---')
+				
+	def ShowWithCandidate(self):
+		print('Show with candidate')
+		for i in range(9):
+			for j in range(9):
+				if self.raws[i][j].is_fixed:
+					fix = 'F'
+				else:
+					fix = ' '
+				print('*{}{}|'.format(self.raws[i][j].value, fix), end='')
+			print('')
+			for jr in range(0,9,3):
+				for j in range(9):
+					str_judge = ''
+					for k in range(3):
+						if self.raws[i][j].judge[jr+k]:
+							str_judge = str_judge + str(jr+k+1)
+						else:
+							str_judge = str_judge + '.'
+					print('{}|'.format(str_judge), end='')
+				print('')
+			if i % 3 == 2:
+				print('')
 
 if __name__ == '__main__':
 	
@@ -156,6 +226,8 @@ if __name__ == '__main__':
 	print('set task')
 	field.SetTask()
 	field.show()
+	#field.ShowWithCandidate()
 	print('solve')
 	field.Solve()
 	field.show()
+	#field.ShowWithCandidate()
